[java] move Options classes into a new options package and deprecate old interface
