--- java/src/main/java/com/saucelabs/saucebindings/InvalidSauceOptionsArgumentException.java
@@ -1,7 +1,0 @@
-package com.saucelabs.saucebindings;
-
-class InvalidSauceOptionsArgumentException extends RuntimeException {
-    public InvalidSauceOptionsArgumentException(String message) {
-        super(message);
-    }
-}
