+++ java/src/main/java/com/saucelabs/saucebindings/examples/BasicOptionsTest.java
@@ -1,5 +1,6 @@
 package com.saucelabs.saucebindings.examples;
 
+import com.saucelabs.saucebindings.options.SauceOptions;
 import com.saucelabs.saucebindings.*;
 import org.junit.Test;
 import org.openqa.selenium.remote.RemoteWebDriver;
