+++ java/src/main/java/com/saucelabs/saucebindings/options/InvalidSauceOptionsArgumentException.java
@@ -1,0 +1,7 @@
+package com.saucelabs.saucebindings.options;
+
+public class InvalidSauceOptionsArgumentException extends RuntimeException {
+    public InvalidSauceOptionsArgumentException(String message) {
+        super(message);
+    }
+}
